requirements.txt
numpy==1.19.5
pandas==1.2.0
python-dateutil==2.8.1
pytz==2020.5
six==1.15.0
